+++ src/com/milanesachan/DRPGTest1/bot/core/CommandManager.java
@@ -10,6 +10,7 @@ import com.milanesachan.DRPGTest1.bot.handlers.guild.GuildDeletionHandler;
 import com.milanesachan.DRPGTest1.bot.handlers.guild.InfoGuildHandler;
 import com.milanesachan.DRPGTest1.bot.handlers.inventory.*;
 import com.milanesachan.DRPGTest1.bot.handlers.item.InfoItemHandler;
+import com.milanesachan.DRPGTest1.bot.handlers.statcards.FusionHandler;
 import com.milanesachan.DRPGTest1.commons.console.ConsoleManager;
 import com.milanesachan.DRPGTest1.networking.DatabaseConnector;
 import net.dv8tion.jda.api.entities.User;
@@ -81,7 +82,7 @@ public class CommandManager extends ListenerAdapter {
         }else if(matchCommand(args[0], "createitem")){
             if(args.length!=3
             || event.getMessage().getMentionedMembers().isEmpty()) {
-                event.getChannel().sendMessage("<@" + userID + "> **Error:** correct format is '>createinv <@User> <itemID>'").queue();
+                event.getChannel().sendMessage("<@" + userID + "> **Error:** correct format is '>createitem <@User> <itemID>'").queue();
             }else{
                 User user = event.getMessage().getMentionedMembers().get(0).getUser();
                 String itemID = args[2];
@@ -99,6 +100,12 @@ public class CommandManager extends ListenerAdapter {
         }else if(matchCommand(args[0], "unequip")){
             UnequipHandler h = new UnequipHandler(event.getChannel(), userID);
             onCharacterRequiredCommand(h, event, userID);
+        }else if(matchCommand(args[0], "fuse")){
+            FusionHandler h = new FusionHandler(event.getChannel(), userID);
+
+            HandlerFilter filter = new HandlerFilter();
+            filter.setCharacterRequired(true);
+            filter.filterHandler(event, h, userID);
         }
     }