+++ src/com/milanesachan/DRPGTest1/commons/parameters/StatCardsParameters.java
@@ -1,0 +1,7 @@
+package com.milanesachan.DRPGTest1.commons.parameters;
+
+
+public class StatCardsParameters {
+    public static float statcardFuseStatMin = 0.2f;
+    public static float statcardFuseStatMax = 0.4f;
+}