+++ src/com/milanesachan/DRPGTest1/game/model/Equipment.java
@@ -5,6 +5,7 @@ import com.milanesachan.DRPGTest1.bot.core.ItemFactory;
 import com.milanesachan.DRPGTest1.commons.exceptions.CharacterNotFoundException;
 import com.milanesachan.DRPGTest1.commons.exceptions.EquipmentNotFoundException;
 import com.milanesachan.DRPGTest1.commons.exceptions.ItemNotFoundException;
+import com.milanesachan.DRPGTest1.game.model.items.StatCard;
 import com.milanesachan.DRPGTest1.game.model.items.Weapon;
 import com.milanesachan.DRPGTest1.networking.DatabaseConnector;
 import net.dv8tion.jda.api.EmbedBuilder;
@@ -18,6 +19,7 @@ import java.sql.Statement;
 public class Equipment implements Embeddable{
     private long userID;
     private Weapon weapon;
+    private StatCard statCard;
 
     public Equipment(long userID){
         this.userID = userID;
@@ -31,6 +33,14 @@ public class Equipment implements Embeddable{
         this.weapon = weapon;
     }
 
+    public StatCard getStatCard() {
+        return statCard;
+    }
+
+    public void setStatCard(StatCard statCard) {
+        this.statCard = statCard;
+    }
+
     public void loadFromDatabase() throws SQLException, EquipmentNotFoundException {
         Connection con = DatabaseConnector.getInstance().getDatabaseConnection();
         if(con != null){
@@ -38,8 +48,14 @@ public class Equipment implements Embeddable{
             ResultSet rs = stmt.executeQuery("SELECT * FROM `equipment` WHERE `UID`="+userID);
             if(rs.next()){
                 String weaponID = rs.getString("WeaponID");
+                String statCardID = rs.getString("StatCardID");
                 try {
                     this.weapon = (Weapon) ItemFactory.getInstance().getItemFromID(weaponID);
+                } catch (ItemNotFoundException e) {
+                    //e.printStackTrace();
+                }
+                try {
+                    this.statCard = (StatCard) ItemFactory.getInstance().getItemFromID(statCardID);
                 } catch (ItemNotFoundException e) {
                     //e.printStackTrace();
                 }
@@ -55,11 +71,12 @@ public class Equipment implements Embeddable{
         if(con != null){
             Statement stmt = con.createStatement();
             String weaponID = weapon != null ? weapon.getItemID() : "";
+            String statCardID = statCard != null ? statCard.getItemID() : "";
             stmt.executeQuery("SELECT * FROM `equipment` WHERE `UID`="+userID);
             if(!stmt.getResultSet().next()){
-                stmt.execute("INSERT INTO `equipment` (UID, WeaponID) VALUES ("+userID+", '"+weaponID+"');");
+                stmt.execute("INSERT INTO `equipment` (UID, WeaponID, StatCardID) VALUES ("+userID+", '"+weaponID+"', '"+statCardID+"');");
             }else{
-                stmt.execute("UPDATE `equipment` SET `WeaponID`='"+weaponID+"' WHERE `UID`="+userID);
+                stmt.execute("UPDATE `equipment` SET `WeaponID`='"+weaponID+"', `StatCardID`='"+statCardID+"' WHERE `UID`="+userID);
             }
             con.close();
         }
@@ -80,7 +97,8 @@ public class Equipment implements Embeddable{
         emb.setColor(0x450000);
         emb.setTitle(charName+"'s Equipment");
         emb.setDescription("User: "+userName);
-        emb.addField("Weapon", weapon != null ? weapon.getItemName() : "No weapon equipped", false);
+        emb.addField("Weapon", weapon != null ? weapon.getItemName() : "*No weapon equipped*", false);
+        emb.addField("Stat Card", statCard != null ? statCard.toString() : "*No stat card equipped*", false);
         emb.setThumbnail(userAvatar);
         return emb;
     }